+++ Sources/StreamChat/Config/StreamRuntimeCheck.swift
@@ -38,4 +38,9 @@ public enum StreamRuntimeCheck {
     ///
     ///  Enables using our legacy web socket connection.
     public static var _useLegacyWebSocketConnection = false
+    
+    /// For *internal use* only
+    ///
+    /// Enables reusing unchanged converted items in database observers.
+    public static var _isDatabaseObserverItemReusingEnabled = true
 }